LayoutTests/http/tests/performance/performance-resource-timing-initiator-no-override-expected.txt

PASS initiator is "img"
PASS foundResources is 1
